Change logo - modify READMEs
Initial PR to update the Logos in the README files to not point to one displaying ForgeRock.

Updated READMEs to be more consistent, added a section for OIDC Login, and added a blurb at the top of each read me about embedded versus OIDC as this is a question from the field.